changed package names and UI improvements
