+++ lib/main.dart
@@ -1,12 +1,14 @@
-import 'package:che/core/theme_color.dart';
-import 'package:che/presentation/mainpage/screen_root.dart';
+import 'presentation/mainpage/screen_root.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'core/theme_color.dart';
 
 void main() {
   runApp(const MyApp());
   SystemChrome.setSystemUIOverlayStyle(const SystemUiOverlayStyle(
+      systemNavigationBarColor: Colors.transparent,
+      systemNavigationBarIconBrightness: Brightness.dark,
       statusBarColor: Colors.transparent,
       statusBarIconBrightness: Brightness.dark));
 }
@@ -19,7 +21,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      title: 'CHE',
+      title: 'Nerve',
       theme: ThemeData(
         scaffoldBackgroundColor: ThemeColor.scaffoldBgColor,
         fontFamily: GoogleFonts.montserrat().fontFamily,
