+++ lib/presentation/home/2015/screen_rev2015.dart
@@ -1,27 +1,28 @@
-import 'package:che/core/const.dart';
-import 'package:che/core/theme_color.dart';
-import 'package:che/presentation/2021/syllabus/screen_syllabus_rev21.dart';
 import 'package:flutter/material.dart';
 
-class ScreenRev2021 extends StatelessWidget {
-  const ScreenRev2021({super.key});
+import '../../../core/const.dart';
+import '../../../core/theme_color.dart';
+import 'lab_manual.dart/screen_lab_manual_rev15.dart';
+import 'syllabus/screen_syllabus_rev15.dart';
+
+class ScreenRev2015 extends StatelessWidget {
+  const ScreenRev2015({super.key});
 
   @override
   Widget build(BuildContext context) {
     final size = MediaQuery.of(context).size;
     return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 30),
+      padding: const EdgeInsets.all(20),
       child: SingleChildScrollView(
         child: Column(
           children: [
-            kHeight20,
             Row(
-              mainAxisAlignment: MainAxisAlignment.spaceAround,
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
                 InkWell(
                   child: Container(
-                    width: size.width * 0.45,
-                    height: size.height * 0.25,
+                    width: size.width * 0.42,
+                    height: size.width * 0.42,
                     decoration: BoxDecoration(
                         boxShadow: const [
                           BoxShadow(
@@ -30,7 +31,7 @@ class ScreenRev2021 extends StatelessWidget {
                               spreadRadius: 0.1,
                               offset: Offset(0, 10)),
                         ],
-                        color: ThemeColor.lightBlue,
+                        color: ThemeColor.white,
                         borderRadius: BorderRadius.circular(10)),
                     padding: const EdgeInsets.all(20),
                     child: Column(
@@ -38,13 +39,133 @@ class ScreenRev2021 extends StatelessWidget {
                       children: const [
                         Icon(
                           Icons.article,
-                          color: ThemeColor.white,
+                          color: ThemeColor.black,
                           size: 30,
                         ),
                         Text(
                           'Syllabus',
                           style: TextStyle(
-                            color: ThemeColor.white,
+                            color: ThemeColor.black,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                  onTap: () {
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (context) => const ScreenSyllabusRev15(
+                          text: 'Syllabus',
+                        ),
+                      ),
+                    );
+                  },
+                ),
+                InkWell(
+                  child: Container(
+                    width: size.width * 0.42,
+                    height: size.width * 0.42,
+                    decoration: BoxDecoration(
+                        boxShadow: const [
+                          BoxShadow(
+                              color: ThemeColor.shadow,
+                              blurRadius: 10,
+                              spreadRadius: 0.1,
+                              offset: Offset(0, 10)),
+                        ],
+                        color: ThemeColor.white,
+                        borderRadius: BorderRadius.circular(10)),
+                    padding: const EdgeInsets.all(20),
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: const [
+                        Icon(
+                          Icons.auto_stories_outlined,
+                          color: ThemeColor.black,
+                          size: 30,
+                        ),
+                        Text(
+                          'Notes',
+                          style: TextStyle(
+                            color: ThemeColor.black,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                  onTap: () {},
+                ),
+              ],
+            ),
+            kHeight25,
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                InkWell(
+                  child: Container(
+                    width: size.width * 0.42,
+                    height: size.width * 0.42,
+                    decoration: BoxDecoration(
+                        boxShadow: const [
+                          BoxShadow(
+                              color: ThemeColor.shadow,
+                              blurRadius: 10,
+                              spreadRadius: 0.1,
+                              offset: Offset(0, 10)),
+                        ],
+                        color: ThemeColor.white,
+                        borderRadius: BorderRadius.circular(10)),
+                    padding: const EdgeInsets.all(20),
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: const [
+                        Icon(
+                          Icons.quiz,
+                          color: ThemeColor.black,
+                          size: 30,
+                        ),
+                        Text(
+                          'Question Paper',
+                          style: TextStyle(
+                            color: ThemeColor.black,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                  onTap: () {},
+                ),
+                InkWell(
+                  child: Container(
+                    width: size.width * 0.42,
+                    height: size.width * 0.42,
+                    decoration: BoxDecoration(
+                        boxShadow: const [
+                          BoxShadow(
+                              color: ThemeColor.shadow,
+                              blurRadius: 10,
+                              spreadRadius: 0.1,
+                              offset: Offset(0, 10)),
+                        ],
+                        color: ThemeColor.white,
+                        borderRadius: BorderRadius.circular(10)),
+                    padding: const EdgeInsets.all(20),
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: const [
+                        Icon(
+                          Icons.menu_book_rounded,
+                          color: ThemeColor.black,
+                          size: 30,
+                        ),
+                        Text(
+                          'Lab Manual',
+                          style: TextStyle(
+                            color: ThemeColor.black,
                             fontWeight: FontWeight.bold,
                           ),
                         ),
@@ -56,184 +177,13 @@ class ScreenRev2021 extends StatelessWidget {
                       context,
                       MaterialPageRoute(
                         builder: (context) =>
-                            const ScreenSyllabusRev21(text: 'Syllabus'),
+                            const ScreenLabManualRev15(text: 'Lab Manual'),
                       ),
                     );
                   },
                 ),
-                InkWell(
-                  child: Container(
-                    width: size.width * 0.45,
-                    height: size.height * 0.25,
-                    decoration: BoxDecoration(
-                        boxShadow: const [
-                          BoxShadow(
-                              color: ThemeColor.shadow,
-                              blurRadius: 10,
-                              spreadRadius: 0.1,
-                              offset: Offset(0, 10)),
-                        ],
-                        color: ThemeColor.lightBlue,
-                        borderRadius: BorderRadius.circular(10)),
-                    padding: const EdgeInsets.all(20),
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: const [
-                        Icon(
-                          Icons.auto_stories_outlined,
-                          color: ThemeColor.white,
-                          size: 30,
-                        ),
-                        Text(
-                          'Notes',
-                          style: TextStyle(
-                            color: ThemeColor.white,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  onTap: () {
-                    // Navigator.push(
-                    //     context,
-                    //     MaterialPageRoute(
-                    //         builder: (context) => )));
-                  },
-                ),
               ],
             ),
-            kHeight25,
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceAround,
-              children: [
-                InkWell(
-                  child: Container(
-                    width: size.width * 0.45,
-                    height: size.height * 0.25,
-                    decoration: BoxDecoration(
-                        boxShadow: const [
-                          BoxShadow(
-                              color: ThemeColor.shadow,
-                              blurRadius: 10,
-                              spreadRadius: 0.1,
-                              offset: Offset(0, 10)),
-                        ],
-                        color: ThemeColor.lightBlue,
-                        borderRadius: BorderRadius.circular(10)),
-                    padding: const EdgeInsets.all(20),
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: const [
-                        Icon(
-                          Icons.quiz,
-                          color: ThemeColor.white,
-                          size: 30,
-                        ),
-                        Text(
-                          'Question paper',
-                          style: TextStyle(
-                            color: ThemeColor.white,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  onTap: () {
-                    // Navigator.push(
-                    //     context,
-                    //     MaterialPageRoute(
-                    //         builder: (context) => )));
-                  },
-                ),
-                InkWell(
-                  child: Container(
-                    width: size.width * 0.45,
-                    height: size.height * 0.25,
-                    decoration: BoxDecoration(
-                        boxShadow: const [
-                          BoxShadow(
-                              color: ThemeColor.shadow,
-                              blurRadius: 10,
-                              spreadRadius: 0.1,
-                              offset: Offset(0, 10)),
-                        ],
-                        color: ThemeColor.lightBlue,
-                        borderRadius: BorderRadius.circular(10)),
-                    padding: const EdgeInsets.all(20),
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: const [
-                        Icon(
-                          Icons.menu_book,
-                          color: ThemeColor.white,
-                          size: 30,
-                        ),
-                        Text(
-                          'Lab Manual',
-                          style: TextStyle(
-                            color: ThemeColor.white,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  onTap: () {
-                    // Navigator.push(
-                    //     context,
-                    //     MaterialPageRoute(
-                    //         builder: (context) => )));
-                  },
-                ),
-              ],
-            ),
-            kHeight25,
-            // Row(
-            //   mainAxisAlignment: MainAxisAlignment.spaceAround,
-            //   children: [
-            //     InkWell(
-            //       child: Container(
-            //         width: size.width * 0.45,
-            //         height: size.height * 0.25,
-            //         decoration: BoxDecoration(
-            //             boxShadow: const [
-            //               BoxShadow(
-            //                   color: ThemeColor.shadow,
-            //                   blurRadius: 10,
-            //                   spreadRadius: 0.1,
-            //                   offset: Offset(0, 10)),
-            //             ],
-            //             color: ThemeColor.lightBlue,
-            //             borderRadius: BorderRadius.circular(10)),
-            //         padding: const EdgeInsets.all(20),
-            //         child: Column(
-            //           mainAxisAlignment: MainAxisAlignment.center,
-            //           children: const [
-            //             Icon(
-            //               Icons.upload,
-            //               color: ThemeColor.white,
-            //             ),
-            //             Text(
-            //               'Upload',
-            //               style: TextStyle(
-            //                 color: ThemeColor.white,
-            //                 fontWeight: FontWeight.bold,
-            //               ),
-            //             ),
-            //           ],
-            //         ),
-            //       ),
-            //       onTap: () {
-            //         // Navigator.push(
-            //         //     context,
-            //         //     MaterialPageRoute(
-            //         //         builder: (context) => )));
-            //       },
-            //     ),
-            //   ],
-            // ),
           ],
         ),
       ),
