+++ lib/presentation/home/2015/syllabus/screen_syllabus_rev15.dart
@@ -1,10 +1,16 @@
-import 'package:che/core/const.dart';
-import 'package:che/core/theme_color.dart';
-import 'package:che/presentation/widgets/custom_sem_button.dart';
 import 'package:flutter/material.dart';
+import '../../../../core/const.dart';
+import '../../../../core/theme_color.dart';
+import '../../../widgets/custom_sem_button.dart';
+import 'widget/sem_five_rev15.dart';
+import 'widget/sem_four_rev15.dart';
+import 'widget/sem_one_rev15.dart';
+import 'widget/sem_six_rev15.dart';
+import 'widget/sem_three_rev15.dart';
+import 'widget/sem_two_rev15.dart';
 
-class ScreenLabManualRev15 extends StatelessWidget {
-  const ScreenLabManualRev15({
+class ScreenSyllabusRev15 extends StatelessWidget {
+  const ScreenSyllabusRev15({
     super.key,
     required this.text,
   });
@@ -15,7 +21,7 @@ class ScreenLabManualRev15 extends StatelessWidget {
     return Scaffold(
       body: SafeArea(
         child: Padding(
-          padding: const EdgeInsets.all(10.0),
+          padding: const EdgeInsets.all(20.0),
           child: Column(
             children: [
               Row(
@@ -46,88 +52,88 @@ class ScreenLabManualRev15 extends StatelessWidget {
                   )
                 ],
               ),
-              kHeight50,
+              kHeight20,
               CustomSemButton(
                 size: size,
                 text: 'Semester I',
                 onTap: () {
-                  // Navigator.push(
-                  //   context,
-                  //   MaterialPageRoute(
-                  //     builder: (context) =>
-                  //         const SemOneSyllabusRev15(text: 'Semester I'),
-                  //   ),
-                  // );
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) =>
+                          const SemOneSyllabusRev15(text: 'Semester I'),
+                    ),
+                  );
                 },
               ),
-              kHeight20,
+              kHeight,
               CustomSemButton(
                 size: size,
                 text: 'Semester II',
                 onTap: () {
-                  // Navigator.push(
-                  //   context,
-                  //   MaterialPageRoute(
-                  //     builder: (context) =>
-                  //         const SemTwoSyllabusRev15(text: 'Semester II'),
-                  //   ),
-                  // );
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) =>
+                          const SemTwoSyllabusRev15(text: 'Semester II'),
+                    ),
+                  );
                 },
               ),
-              kHeight20,
+              kHeight,
               CustomSemButton(
                 size: size,
                 text: 'Semester III',
                 onTap: () {
-                  // Navigator.push(
-                  //   context,
-                  //   MaterialPageRoute(
-                  //     builder: (context) =>
-                  //         const SemThreeSyllabusRev15(text: 'Semester III'),
-                  //   ),
-                  // );
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) =>
+                          const SemThreeSyllabusRev15(text: 'Semester III'),
+                    ),
+                  );
                 },
               ),
-              kHeight20,
+              kHeight,
               CustomSemButton(
                 size: size,
                 text: 'Semester IV',
                 onTap: () {
-                  // Navigator.push(
-                  //   context,
-                  //   MaterialPageRoute(
-                  //     builder: (context) =>
-                  //         const SemFourSyllabusRev15(text: 'Semester IV'),
-                  //   ),
-                  // );
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) =>
+                          const SemFourSyllabusRev15(text: 'Semester IV'),
+                    ),
+                  );
                 },
               ),
-              kHeight20,
+              kHeight,
               CustomSemButton(
                 size: size,
                 text: 'Semester V',
                 onTap: () {
-                  // Navigator.push(
-                  //   context,
-                  //   MaterialPageRoute(
-                  //     builder: (context) =>
-                  //         const SemFiveSyllabusRev15(text: 'Semester V'),
-                  //   ),
-                  // );
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) =>
+                          const SemFiveSyllabusRev15(text: 'Semester V'),
+                    ),
+                  );
                 },
               ),
-              kHeight20,
+              kHeight,
               CustomSemButton(
                 size: size,
                 text: 'Semester VI',
                 onTap: () {
-                  // Navigator.push(
-                  //   context,
-                  //   MaterialPageRoute(
-                  //     builder: (context) =>
-                  //         const SemSixSyllabusRev15(text: 'Semester VI'),
-                  //   ),
-                  // );
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) =>
+                          const SemSixSyllabusRev15(text: 'Semester VI'),
+                    ),
+                  );
                 },
               ),
             ],
