+++ lib/presentation/home/2015/syllabus/widget/sem_two_rev15.dart
@@ -1,0 +1,247 @@
+import 'package:flutter/material.dart';
+
+import '../../../../../core/const.dart';
+import '../../../../../core/pdf/syllabus/syllabus15_pdf_path.dart';
+import '../../../../../core/theme_color.dart';
+import '../../../../widgets/custom_sub_button.dart';
+import '../../../../widgets/pdf_viewer.dart';
+
+class SemTwoSyllabusRev15 extends StatelessWidget {
+  const SemTwoSyllabusRev15({
+    super.key,
+    required this.text,
+  });
+  final String text;
+  @override
+  Widget build(BuildContext context) {
+    final size = MediaQuery.of(context).size;
+    return Scaffold(
+      body: SafeArea(
+        child: Padding(
+          padding: const EdgeInsets.all(20.0),
+          child: Column(
+            children: [
+              Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  Text(
+                    text,
+                    style: const TextStyle(
+                      color: ThemeColor.black,
+                      fontSize: 26,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                  InkWell(
+                    child: Container(
+                      decoration: BoxDecoration(
+                          color: ThemeColor.lightGrey,
+                          borderRadius: BorderRadius.circular(12)),
+                      padding: const EdgeInsets.all(12),
+                      child: const Icon(
+                        Icons.arrow_back,
+                        color: Colors.white,
+                      ),
+                    ),
+                    onTap: () {
+                      Navigator.pop(context);
+                    },
+                  )
+                ],
+              ),
+              kHeight20,
+              Expanded(
+                child: SingleChildScrollView(
+                  child: Column(
+                    children: [
+                      InkWell(
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => const PdfViewer(
+                                src: Syllabus15PdfPath.english2,
+                              ),
+                            ),
+                          );
+                        },
+                        child: CustomSubButton(
+                          size: size,
+                          coursename: 'English for Communication II',
+                          coursecode: '2001',
+                        ),
+                      ),
+                      kHeight,
+                      InkWell(
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => const PdfViewer(
+                                src: Syllabus15PdfPath.maths2,
+                              ),
+                            ),
+                          );
+                        },
+                        child: CustomSubButton(
+                          size: size,
+                          coursename: 'Engineering Mathematics II',
+                          coursecode: '2002',
+                        ),
+                      ),
+                      kHeight,
+                      InkWell(
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => const PdfViewer(
+                                src: Syllabus15PdfPath.physics2,
+                              ),
+                            ),
+                          );
+                        },
+                        child: CustomSubButton(
+                          size: size,
+                          coursename: 'Engineering Physics II',
+                          coursecode: '2003',
+                        ),
+                      ),
+                      kHeight,
+                      InkWell(
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => const PdfViewer(
+                                src: Syllabus15PdfPath.chemistry2,
+                              ),
+                            ),
+                          );
+                        },
+                        child: CustomSubButton(
+                          size: size,
+                          coursename: 'Engineering Chemistry II',
+                          coursecode: '2004',
+                        ),
+                      ),
+                      kHeight,
+                      InkWell(
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => const PdfViewer(
+                                src: Syllabus15PdfPath.graphics2,
+                              ),
+                            ),
+                          );
+                        },
+                        child: CustomSubButton(
+                          size: size,
+                          coursename: 'Engineering Graphics',
+                          coursecode: '2005',
+                        ),
+                      ),
+                      kHeight,
+                      InkWell(
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => const PdfViewer(
+                                src: Syllabus15PdfPath.programingc,
+                              ),
+                            ),
+                          );
+                        },
+                        child: CustomSubButton(
+                          size: size,
+                          coursename: 'Programming in C',
+                          coursecode: '2131',
+                        ),
+                      ),
+                      kHeight,
+                      InkWell(
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => const PdfViewer(
+                                src: Syllabus15PdfPath.workshop,
+                              ),
+                            ),
+                          );
+                        },
+                        child: CustomSubButton(
+                          size: size,
+                          coursename: 'Workshop Practice',
+                          coursecode: '2008',
+                        ),
+                      ),
+                      kHeight,
+                      InkWell(
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => const PdfViewer(
+                                src: Syllabus15PdfPath.lifeskill,
+                              ),
+                            ),
+                          );
+                        },
+                        child: CustomSubButton(
+                          size: size,
+                          coursename: 'Life Skill',
+                          coursecode: '2009',
+                        ),
+                      ),
+                      kHeight,
+                      InkWell(
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => const PdfViewer(
+                                src: Syllabus15PdfPath.sciencelab,
+                              ),
+                            ),
+                          );
+                        },
+                        child: CustomSubButton(
+                          size: size,
+                          coursename: 'Engineering Science Lab I',
+                          coursecode: '2007',
+                        ),
+                      ),
+                      kHeight,
+                      InkWell(
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => const PdfViewer(
+                                src: Syllabus15PdfPath.clab,
+                              ),
+                            ),
+                          );
+                        },
+                        child: CustomSubButton(
+                          size: size,
+                          coursename: 'Programming in C Lab',
+                          coursecode: '2139	',
+                        ),
+                      ),
+                      kHeight,
+                    ],
+                  ),
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
