+++ lib/presentation/home/2021/syllabus/screen_syllabus_rev21.dart
@@ -1,16 +1,15 @@
-import 'package:che/core/const.dart';
-import 'package:che/core/theme_color.dart';
-import 'package:che/presentation/2015/syllabus/widget/sem_five_rev15.dart';
-import 'package:che/presentation/2015/syllabus/widget/sem_four_rev15.dart';
-import 'package:che/presentation/2015/syllabus/widget/sem_one_rev15.dart';
-import 'package:che/presentation/2015/syllabus/widget/sem_six_rev15.dart';
-import 'package:che/presentation/2015/syllabus/widget/sem_three_rev15.dart';
-import 'package:che/presentation/2015/syllabus/widget/sem_two_rev15.dart';
-import 'package:che/presentation/widgets/custom_sem_button.dart';
 import 'package:flutter/material.dart';
 
-class ScreenSyllabusRev15 extends StatelessWidget {
-  const ScreenSyllabusRev15({
+import '../../../../core/const.dart';
+import '../../../../core/theme_color.dart';
+import '../../../widgets/custom_sem_button.dart';
+import 'widgets/sem_four_rev21.dart';
+import 'widgets/sem_one_rev21.dart';
+import 'widgets/sem_three_rev21.dart';
+import 'widgets/sem_two_rev21.dart';
+
+class ScreenSyllabusRev21 extends StatelessWidget {
+  const ScreenSyllabusRev21({
     super.key,
     required this.text,
   });
@@ -21,7 +20,7 @@ class ScreenSyllabusRev15 extends StatelessWidget {
     return Scaffold(
       body: SafeArea(
         child: Padding(
-          padding: const EdgeInsets.all(10.0),
+          padding: const EdgeInsets.all(20.0),
           child: Column(
             children: [
               Row(
@@ -52,7 +51,7 @@ class ScreenSyllabusRev15 extends StatelessWidget {
                   )
                 ],
               ),
-              kHeight50,
+              kHeight20,
               CustomSemButton(
                 size: size,
                 text: 'Semester I',
@@ -61,12 +60,12 @@ class ScreenSyllabusRev15 extends StatelessWidget {
                     context,
                     MaterialPageRoute(
                       builder: (context) =>
-                          const SemOneSyllabusRev15(text: 'Semester I'),
+                          const SemOneSyllabusRev21(text: 'Semester I'),
                     ),
                   );
                 },
               ),
-              kHeight20,
+              kHeight,
               CustomSemButton(
                 size: size,
                 text: 'Semester II',
@@ -75,12 +74,12 @@ class ScreenSyllabusRev15 extends StatelessWidget {
                     context,
                     MaterialPageRoute(
                       builder: (context) =>
-                          const SemTwoSyllabusRev15(text: 'Semester II'),
+                          const SemTwoSyllabusRev21(text: 'Semester II'),
                     ),
                   );
                 },
               ),
-              kHeight20,
+              kHeight,
               CustomSemButton(
                 size: size,
                 text: 'Semester III',
@@ -89,12 +88,12 @@ class ScreenSyllabusRev15 extends StatelessWidget {
                     context,
                     MaterialPageRoute(
                       builder: (context) =>
-                          const SemThreeSyllabusRev15(text: 'Semester III'),
+                          const SemThreeSyllabusRev21(text: 'Semester III'),
                     ),
                   );
                 },
               ),
-              kHeight20,
+              kHeight,
               CustomSemButton(
                 size: size,
                 text: 'Semester IV',
@@ -103,39 +102,39 @@ class ScreenSyllabusRev15 extends StatelessWidget {
                     context,
                     MaterialPageRoute(
                       builder: (context) =>
-                          const SemFourSyllabusRev15(text: 'Semester IV'),
+                          const SemFourSyllabusRev21(text: 'Semester IV'),
                     ),
                   );
                 },
               ),
-              kHeight20,
-              CustomSemButton(
-                size: size,
-                text: 'Semester V',
-                onTap: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) =>
-                          const SemFiveSyllabusRev15(text: 'Semester V'),
-                    ),
-                  );
-                },
-              ),
-              kHeight20,
-              CustomSemButton(
-                size: size,
-                text: 'Semester VI',
-                onTap: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) =>
-                          const SemSixSyllabusRev15(text: 'Semester VI'),
-                    ),
-                  );
-                },
-              ),
+              kHeight,
+              // CustomSemButton(
+              //   size: size,
+              //   text: 'Semester V',
+              //   onTap: () {
+              //     // Navigator.push(
+              //     //   context,
+              //     //   MaterialPageRoute(
+              //     //     builder: (context) =>
+              //     //         const SemFiveSyllabusRev15(text: 'Semester V'),
+              //     //   ),
+              //     // );
+              //   },
+              // ),
+              // kHeight,
+              // CustomSemButton(
+              //   size: size,
+              //   text: 'Semester VI',
+              //   onTap: () {
+              //     // Navigator.push(
+              //     //   context,
+              //     //   MaterialPageRoute(
+              //     //     builder: (context) =>
+              //     //         const SemSixSyllabusRev15(text: 'Semester VI'),
+              //     //   ),
+              //     // );
+              //   },
+              // ),
             ],
           ),
         ),
