+++ lib/presentation/home/screen_home.dart
@@ -1,9 +1,11 @@
 import 'package:animated_segmented_tab_control/animated_segmented_tab_control.dart';
-import 'package:che/core/const.dart';
-import 'package:che/core/theme_color.dart';
-import 'package:che/presentation/2015/screen_rev2015.dart';
-import 'package:che/presentation/2021/screen_rev2021.dart';
+
 import 'package:flutter/material.dart';
+
+import '../../core/const.dart';
+import '../../core/theme_color.dart';
+import '2015/screen_rev2015.dart';
+import '2021/screen_rev2021.dart';
 
 class ScreenHome extends StatelessWidget {
   const ScreenHome({super.key});
@@ -14,18 +16,18 @@ class ScreenHome extends StatelessWidget {
       length: 2,
       child: Scaffold(
         body: SafeArea(
-          child: Padding(
-            padding: const EdgeInsets.symmetric(vertical: 10),
-            child: Column(
-              children: [
-                kHeight,
-                Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 10),
+          child: Column(
+            children: [
+              kHeight,
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 10),
+                child: Padding(
+                  padding: const EdgeInsets.all(20.0),
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.start,
                     children: const [
                       Text(
-                        'Study struggle',
+                        'Nerve',
                         style: TextStyle(
                           color: ThemeColor.black,
                           fontSize: 30,
@@ -35,39 +37,38 @@ class ScreenHome extends StatelessWidget {
                     ],
                   ),
                 ),
-                kHeight30,
-                const Padding(
-                  padding: EdgeInsets.symmetric(horizontal: 10),
-                  child: SegmentedTabControl(
-                    height: 50,
-                    radius: Radius.circular(10),
-                    backgroundColor: ThemeColor.lightGrey,
-                    indicatorColor: ThemeColor.primary,
-                    tabTextColor: ThemeColor.black,
-                    selectedTabTextColor: ThemeColor.white,
-                    textStyle: TextStyle(
-                      fontWeight: FontWeight.w500,
+              ),
+              const Padding(
+                padding: EdgeInsets.only(left: 20.0, right: 20.0),
+                child: SegmentedTabControl(
+                  height: 40,
+                  radius: Radius.circular(10),
+                  backgroundColor: ThemeColor.lightGrey,
+                  indicatorColor: ThemeColor.primary,
+                  tabTextColor: ThemeColor.black,
+                  selectedTabTextColor: ThemeColor.white,
+                  textStyle: TextStyle(
+                    fontWeight: FontWeight.w500,
+                  ),
+                  tabs: [
+                    SegmentTab(
+                      label: "Rev 2015",
                     ),
-                    tabs: [
-                      SegmentTab(
-                        label: "Rev 2015",
-                      ),
-                      SegmentTab(
-                        label: "Rev 2021",
-                      ),
-                    ],
-                  ),
+                    SegmentTab(
+                      label: "Rev 2021",
+                    ),
+                  ],
                 ),
-                const Expanded(
-                  child: TabBarView(
-                    children: [
-                      ScreenRev2015(),
-                      ScreenRev2021(),
-                    ],
-                  ),
-                )
-              ],
-            ),
+              ),
+              const Expanded(
+                child: TabBarView(
+                  children: [
+                    ScreenRev2015(),
+                    ScreenRev2021(),
+                  ],
+                ),
+              )
+            ],
           ),
         ),
       ),
