+++ lib/presentation/mainpage/screen_root.dart
@@ -1,7 +1,9 @@
-import 'package:che/presentation/about/screen_about.dart';
-import 'package:che/presentation/home/screen_home.dart';
-import 'package:che/presentation/mainpage/widget/bottom_navbar.dart';
+
 import 'package:flutter/material.dart';
+
+import '../about/screen_about.dart';
+import '../home/screen_home.dart';
+import 'widget/bottom_navbar.dart';
 
 class ScreenRoot extends StatelessWidget {
   ScreenRoot({super.key});
