+++ lib/presentation/widgets/custom_sem_button.dart
@@ -1,4 +1,4 @@
-import 'package:che/core/theme_color.dart';
+import '../../core/theme_color.dart';
 import 'package:flutter/material.dart';
 
 class CustomSemButton extends StatelessWidget {
